moved profiling trace folder creation logic in Sentry class moved profilingTracesDirPath from SentryAndroidOptions to SentryOptions
