+++ sentry-android-core/src/main/java/io/sentry/android/core/AndroidOptionsInitializer.java
@@ -14,7 +14,6 @@ import io.sentry.SendFireAndForgetOutboxSender;
 import io.sentry.SentryLevel;
 import io.sentry.android.fragment.FragmentLifecycleIntegration;
 import io.sentry.android.timber.SentryTimberIntegration;
-import io.sentry.util.FileUtils;
 import io.sentry.util.Objects;
 import java.io.BufferedInputStream;
 import java.io.File;
@@ -297,31 +296,12 @@ final class AndroidOptionsInitializer {
    * @param context the Application context
    * @param options the SentryAndroidOptions
    */
-  @SuppressWarnings("FutureReturnValueIgnored")
   private static void initializeCacheDirs(
       final @NotNull Context context, final @NotNull SentryAndroidOptions options) {
     final File cacheDir = new File(context.getCacheDir(), "sentry");
     final File profilingTracesDir = new File(cacheDir, "profiling_traces");
     options.setCacheDirPath(cacheDir.getAbsolutePath());
     options.setProfilingTracesDirPath(profilingTracesDir.getAbsolutePath());
-
-    // We don't have the options set by code (manual initialisation) at this point, so we have no
-    // way to know if profiling is manually enabled or not.
-    // In case it's disabled the overhead should be low enough
-    profilingTracesDir.mkdirs();
-    final File[] oldTracesDirContent = profilingTracesDir.listFiles();
-
-    options
-        .getExecutorService()
-        .submit(
-            () -> {
-              if (oldTracesDirContent == null) return;
-              // Method trace files are normally deleted at the end of traces, but if that fails
-              // for some reason we try to clear any old files here.
-              for (File f : oldTracesDirContent) {
-                FileUtils.deleteRecursively(f);
-              }
-            });
   }
 
   private static boolean isNdkAvailable(final @NotNull BuildInfoProvider buildInfoProvider) {
