+++ sentry-android-core/src/main/java/io/sentry/android/core/SentryAndroidOptions.java
@@ -7,7 +7,6 @@ import io.sentry.SentryOptions;
 import io.sentry.SpanStatus;
 import io.sentry.protocol.SdkVersion;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 /** Sentry SDK options for Android */
 public final class SentryAndroidOptions extends SentryOptions {
@@ -84,9 +83,6 @@ public final class SentryAndroidOptions extends SentryOptions {
    * Spans.
    */
   private boolean enableActivityLifecycleTracingAutoFinish = true;
-
-  /** The cache dir. path for caching profiling traces */
-  private @Nullable String profilingTracesDirPath;
 
   /** Interval for profiling traces in milliseconds. Defaults to 100 times per second */
   private int profilingTracesIntervalMillis = 1_000 / 100;
@@ -224,24 +220,6 @@ public final class SentryAndroidOptions extends SentryOptions {
     enableSystemEventBreadcrumbs = enable;
     enableAppLifecycleBreadcrumbs = enable;
     enableUserInteractionBreadcrumbs = enable;
-  }
-
-  /**
-   * Returns the profiling traces dir. path if set
-   *
-   * @return the profiling traces dir. path or null if not set
-   */
-  public @Nullable String getProfilingTracesDirPath() {
-    return profilingTracesDirPath;
-  }
-
-  /**
-   * Sets the profiling traces dir. path
-   *
-   * @param profilingTracesDirPath the profiling traces dir. path
-   */
-  public void setProfilingTracesDirPath(@Nullable String profilingTracesDirPath) {
-    this.profilingTracesDirPath = profilingTracesDirPath;
   }
 
   /**
